+++ x/btcstaking/keeper/keeper.go
@@ -61,13 +61,13 @@ func (k Keeper) Logger(ctx sdk.Context) log.Logger {
 // BeginBlocker is invoked upon `BeginBlock` of the system. The function
 // iterates over all BTC delegations under non-slashed finality providers
 // to 1) record the voting power table for the current height, and 2) record
-// the voting power distribution cache used for distributing rewards once
-// the block is finalised by finality providers.
+// the voting power distribution cache used for computing voting power table
+// and distributing rewards once the block is finalised by finality providers.
 func (k Keeper) BeginBlocker(ctx context.Context) error {
 	// index BTC height at the current height
 	k.IndexBTCHeight(ctx)
 	// update voting power distribution
-	k.UpdatePowerDist(ctx)
+	k.UpdatePowerDist_OLD(ctx)
 
 	return nil
 }
