+++ x/btcstaking/keeper/voting_power_table_test.go
@@ -74,6 +74,7 @@ func FuzzVotingPowerTable(f *testing.F) {
 					slashingChangeLockTime,
 				)
 				require.NoError(t, err)
+				btclcKeeper.EXPECT().GetTipInfo(gomock.Any()).Return(&btclctypes.BTCHeaderInfo{Height: 0}).Times(1)
 				err = keeper.AddBTCDelegation(ctx, btcDel)
 				require.NoError(t, err)
 			}
@@ -243,6 +244,7 @@ func FuzzVotingPowerTable_ActiveFinalityProviders(f *testing.F) {
 				slashingChangeLockTime,
 			)
 			require.NoError(t, err)
+			btclcKeeper.EXPECT().GetTipInfo(gomock.Any()).Return(&btclctypes.BTCHeaderInfo{Height: 1}).Times(1)
 			err = keeper.AddBTCDelegation(ctx, btcDel)
 			require.NoError(t, err)
 
@@ -303,6 +305,7 @@ func FuzzVotingPowerTable_ActiveFinalityProviderRotation(f *testing.F) {
 
 		// mock BTC light client and BTC checkpoint modules
 		btclcKeeper := types.NewMockBTCLightClientKeeper(ctrl)
+		btclcKeeper.EXPECT().GetTipInfo(gomock.Any()).Return(&btclctypes.BTCHeaderInfo{Height: 1}).AnyTimes()
 		btccKeeper := types.NewMockBtcCheckpointKeeper(ctrl)
 		btccKeeper.EXPECT().GetParams(gomock.Any()).Return(btcctypes.DefaultParams()).AnyTimes()
 		keeper, ctx := keepertest.BTCStakingKeeper(t, btclcKeeper, btccKeeper)
@@ -360,7 +363,6 @@ func FuzzVotingPowerTable_ActiveFinalityProviderRotation(f *testing.F) {
 		// record voting power table
 		babylonHeight := datagen.RandomInt(r, 10) + 1
 		ctx = datagen.WithCtxHeight(ctx, babylonHeight)
-		btclcKeeper.EXPECT().GetTipInfo(gomock.Any()).Return(&btclctypes.BTCHeaderInfo{Height: 1}).Times(1)
 		err = keeper.BeginBlocker(ctx)
 		require.NoError(t, err)
 
@@ -405,7 +407,6 @@ func FuzzVotingPowerTable_ActiveFinalityProviderRotation(f *testing.F) {
 		// record voting power table
 		babylonHeight += 1
 		ctx = datagen.WithCtxHeight(ctx, babylonHeight)
-		btclcKeeper.EXPECT().GetTipInfo(gomock.Any()).Return(&btclctypes.BTCHeaderInfo{Height: 1}).Times(1)
 		err = keeper.BeginBlocker(ctx)
 		require.NoError(t, err)
